refactor(persona): fix budget_hit false-positive, harden commits, split file
Address CodeRabbit's re-review of the systemic-guard change:

- `budget_hit` was re-derived in `run` from `budget.exhausted()`, which is true
  whenever a run merely fills its session allowance exactly — dropping nothing.
  Removed: the selection loop already sets `budget_hit` precisely when it drops a
  pending session, and the call-budget path sets it on `BudgetExhausted`, so the
  flag now reflects only a real early stop.
- The deferred-commit loop used `?`, so one failed cursor write aborted the whole
  run and discarded the reduce output (the pack). It now logs and continues per
  key — a cursor is a fast-skip, not a correctness gate, so the session simply
  re-digests next run.
- Split the git-history ingestion (`ingest_git` + its helpers, feature-gated)
  into a sibling `pipeline_git.rs`, bringing `pipeline.rs` back under the repo's
  500-line limit (557 → 492).
- Pinned `max_llm_calls` in the two truncation pipeline tests so they isolate
  loss behavior and don't break if the default call ceiling is lowered later.

Tests: persona 65, persona+git-diff 69 (exercises the moved ingest_git),
all-features and persona matrix green; fmt / default clippy / doc gate clean.
